+++ modules/local/prepare_prediction_input/templates/prepare_prediction_input.py
@@ -138,7 +138,8 @@ class Utils:
                 alleles = [line.strip() for line in f.readlines()]
         else:
             alleles = allele_str.split(";")
-        alleles_normalized = [mhcgnomes.parse(allele).to_string() for allele in alleles]
+        # Truncate 3- and 4-field typings down to 2 fields so higher-resolution HLA inputs match the 2-field entries in assets/supported_alleles.json
+        alleles_normalized = [mhcgnomes.parse(allele).restrict_allele_fields(2).to_string() for allele in alleles]
 
         return alleles_normalized